Make product menu component extend existing menu component
