+++ src/org/labkey/test/components/html/BootstrapMenu.java
@@ -17,28 +17,21 @@ package org.labkey.test.components.html;
 
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
-import org.labkey.test.components.Component;
-import org.labkey.test.components.WebDriverComponent;
 import org.labkey.test.util.LogMethod;
 import org.labkey.test.util.LoggedParam;
-import org.labkey.test.util.TestLogger;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
 import java.util.List;
 
-
-public class BootstrapMenu extends WebDriverComponent<BootstrapMenu.ElementCache>
+public class BootstrapMenu extends BaseBootstrapMenu
 {
-    protected final WebDriver _driver;
-    protected final WebElement _componentElement;
     private int _expandRetryCount = 1;
 
     /* componentElement should contain the toggle anchor *and* the UL containing list items */
     public BootstrapMenu(WebDriver driver, WebElement componentElement)
     {
-        _componentElement = componentElement;
-        _driver = driver;
+        super(driver, componentElement);
     }
 
     static public BootstrapMenuFinder finder(WebDriver driver)
@@ -55,53 +48,10 @@ public class BootstrapMenu extends WebDriverComponent<BootstrapMenu.ElementCache
         return new BootstrapMenuFinder(driver).withButtonTextContaining(menuToggleText).find();
     }
 
-    /* Sometimes the menu doesn't expand on the first try.
-     * Sets the number of attempts it will make to expand the menu before failing/giving up. */
+    @Override
     public BootstrapMenu withExpandRetries(int retries)
     {
-        _expandRetryCount = retries;
-        return this;
-    }
-
-    @Override
-    protected WebDriver getDriver()
-    {
-        return _driver;
-    }
-
-    @Override
-    public WebElement getComponentElement()
-    {
-        return _componentElement;
-    }
-
-    public boolean isExpanded()
-    {
-        return "true".equals(elementCache().toggleAnchor.getAttribute("aria-expanded"));
-    }
-
-    public void expand()
-    {
-        if (!isExpanded())
-        {
-            getWrapper().scrollIntoView(elementCache().toggleAnchor);
-            for (int retry = 0; retry < _expandRetryCount; retry++)
-            {
-                elementCache().toggleAnchor.click();
-                if (WebDriverWrapper.waitFor(this::isExpanded, 1000))
-                    break;
-                else
-                    TestLogger.log("retrying menu expand, attempt #" + retry);
-            }
-        }
-        WebDriverWrapper.waitFor(this::isExpanded, "Menu did not expand as expected", WebDriverWrapper.WAIT_FOR_JAVASCRIPT);
-    }
-
-    public void collapse()
-    {
-        if (isExpanded())
-            elementCache().toggleAnchor.click();
-        WebDriverWrapper.waitFor(()-> !isExpanded(), "Menu did not collapse as expected", WebDriverWrapper.WAIT_FOR_JAVASCRIPT);
+        return (BootstrapMenu) super.withExpandRetries(retries);
     }
 
     public List<WebElement> findVisibleMenuItems()
@@ -165,30 +115,19 @@ public class BootstrapMenu extends WebDriverComponent<BootstrapMenu.ElementCache
     }
 
     @Override
+    protected ElementCache elementCache()
+    {
+        return (ElementCache) super.elementCache();
+    }
+
+    @Override
     protected ElementCache newElementCache()
     {
         return new ElementCache();
     }
 
-    protected class ElementCache extends Component.ElementCache
+    protected class ElementCache extends BaseBootstrapMenu.ElementCache
     {
-        public final WebElement toggleAnchor = getToggleLocator().findWhenNeeded(getComponentElement());
-
-        public WebElement findOpenMenu()
-        {
-            WebElement insideContainerList = Locator.tagWithClassContaining("ul", "dropdown-menu")
-                    .findElementOrNull(getComponentElement());
-            if (insideContainerList != null)
-                return insideContainerList;
-
-            // outside the container, require it to be block-display,
-            // as is the case with dataRegion header menus.
-            return Locator.tagWithClassContaining("ul", "dropdown-menu")
-                    .notHidden()
-                    .withAttributeContaining("style", "display: block")
-                    .findElement(getDriver());
-        }
-
         public WebElement findVisibleMenuPanel()
         {
             WebElement menuList = findOpenMenu();
